TYPESCRIPT to  JavaScript refactor. Init JEST tests
I have decided to give up on working with typescript for now as it consumes a lot of my time on struggling with interating external libraries with typescript.

I descided to work on actual functionality in the first place. In the future I'm going to come back to integrating React app with TS and do more research about it.

Other:
- Configured eslint/prettier
- configured JEST + react-testing-library
- added first unit tests for Notification component